+++ scm-ui/ui-webapp/src/containers/Profile.tsx
@@ -65,6 +65,16 @@ class Profile extends React.Component<Props> {
     return this.stripEndingSlash(this.props.match.url);
   };
 
+  mayChangePassword = () => {
+    const { me } = this.props;
+    return !!me?._links?.password;
+  };
+
+  canManagePublicKeys = () => {
+    const { me } = this.props;
+    return !!me?._links?.publicKeys;
+  };
+
   render() {
     const url = this.matchedUrl();
 
@@ -94,8 +104,12 @@ class Profile extends React.Component<Props> {
           <CustomQueryFlexWrappedColumns>
             <PrimaryContentColumn>
               <Route path={url} exact render={() => <ProfileInfo me={me} />} />
-              <Route path={`${url}/settings/password`} render={() => <ChangeUserPassword me={me} />} />
-              <Route path={`${url}/settings/publicKeys`} render={() => <SetPublicKeys user={me}/>} />
+              {this.mayChangePassword() && (
+                <Route path={`${url}/settings/password`} render={() => <ChangeUserPassword me={me} />} />
+              )}
+              {this.canManagePublicKeys() && (
+                <Route path={`${url}/settings/publicKeys`} render={() => <SetPublicKeys user={me} />} />
+              )}
               <ExtensionPoint name="profile.route" props={extensionProps} renderAll={true} />
             </PrimaryContentColumn>
             <SecondaryNavigationColumn>
@@ -106,15 +120,17 @@ class Profile extends React.Component<Props> {
                   label={t("profile.informationNavLink")}
                   title={t("profile.informationNavLink")}
                 />
-                <SubNavigation
-                  to={`${url}/settings/password`}
-                  label={t("profile.settingsNavLink")}
-                  title={t("profile.settingsNavLink")}
-                >
-                  <NavLink to={`${url}/settings/password`} label={t("profile.changePasswordNavLink")} />
-                  <SetPublicKeyNavLink user={me} publicKeyUrl={`${url}/settings/publicKeys`} />
-                  <ExtensionPoint name="profile.setting" props={extensionProps} renderAll={true} />
-                </SubNavigation>
+                {this.mayChangePassword() && (
+                  <SubNavigation
+                    to={`${url}/settings/password`}
+                    label={t("profile.settingsNavLink")}
+                    title={t("profile.settingsNavLink")}
+                  >
+                    <NavLink to={`${url}/settings/password`} label={t("profile.changePasswordNavLink")} />
+                    <SetPublicKeyNavLink user={me} publicKeyUrl={`${url}/settings/publicKeys`} />
+                    <ExtensionPoint name="profile.setting" props={extensionProps} renderAll={true} />
+                  </SubNavigation>
+                )}
               </SecondaryNavigation>
             </SecondaryNavigationColumn>
           </CustomQueryFlexWrappedColumns>