+++ scm-ui/ui-webapp/src/modules/auth.ts
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-import { Me } from "@scm-manager/ui-types";
+import { Link, Me } from "@scm-manager/ui-types";
 import * as types from "./types";
 
 import { apiClient, UnauthorizedError } from "@scm-manager/ui-components";
@@ -35,6 +35,7 @@ import {
   fetchIndexResourcesSuccess,
   getLoginLink
 } from "./indexResource";
+import { AnyAction } from "redux";
 
 // Action
 
@@ -61,7 +62,7 @@ const initialState = {};
 
 export default function reducer(
   state: object = initialState,
-  action: object = {
+  action: AnyAction = {
     type: "UNKNOWN"
   }
 ) {
@@ -174,7 +175,7 @@ const callFetchMe = (link: string): Promise<Me> => {
 };
 
 export const login = (loginLink: string, username: string, password: string) => {
-  const login_data = {
+  const loginData = {
     cookie: true,
     grant_type: "password",
     username,
@@ -183,14 +184,14 @@ export const login = (loginLink: string, username: string, password: string) => 
   return function(dispatch: any) {
     dispatch(loginPending());
     return apiClient
-      .post(loginLink, login_data)
+      .post(loginLink, loginData)
       .then(() => {
         dispatch(fetchIndexResourcesPending());
         return callFetchIndexResources();
       })
       .then(response => {
         dispatch(fetchIndexResourcesSuccess(response));
-        const meLink = response._links.me.href;
+        const meLink = (response._links.me as Link).href;
         return callFetchMe(meLink);
       })
       .then(me => {
@@ -219,7 +220,7 @@ export const fetchMe = (link: string) => {
   };
 };
 
-export const logout = (link: string) => {
+export const logout = (link: string, callback: () => void) => {
   return function(dispatch: any) {
     dispatch(logoutPending());
     return apiClient
@@ -247,6 +248,7 @@ export const logout = (link: string) => {
           dispatch(fetchIndexResources());
         }
       })
+      .then(callback)
       .catch(error => {
         dispatch(logoutFailure(error));
       });
@@ -256,17 +258,17 @@ export const logout = (link: string) => {
 // selectors
 
 const stateAuth = (state: object): object => {
+  // @ts-ignore Right types for redux not available
   return state.auth || {};
 };
 
 export const isAuthenticated = (state: object) => {
-  if (state.auth.me && !getLoginLink(state)) {
-    return true;
-  }
-  return false;
+  // @ts-ignore Right types for redux not available
+  return !!((state.auth.me && !getLoginLink(state)) || isAnonymous(state.auth.me));
 };
 
 export const getMe = (state: object): Me => {
+  // @ts-ignore Right types for redux not available
   return stateAuth(state).me;
 };
 
@@ -295,5 +297,12 @@ export const getLogoutFailure = (state: object) => {
 };
 
 export const isRedirecting = (state: object) => {
+  // @ts-ignore Right types for redux not available
   return !!stateAuth(state).redirecting;
+};
+
+// Helper methods
+
+export const isAnonymous = (me: Me) => {
+  return me?.name === "_anonymous";
 };